+++ src/utils.c
@@ -1,6 +1,6 @@
 #include "utils.h"
 
-size_t GetElementIndex(int value, int *array, int length) {
+size_t GetElementIndex(int value, int *array, size_t length) {
   for (size_t i = 0; i < length; i++) {
     if (array[i] == value)
       return i;
