+++ src/utils.h
@@ -1,3 +1,3 @@
 #include <stdio.h>
 
-size_t GetElementIndex(int value, int *array, int length);
+size_t GetElementIndex(int value, int *array, size_t length);
